Crate ordered_channel

Equivalent of mpsc or crossbeam-channel, but allows sending messages with an index associated with them, to be delivered in the specific indexed order.

This enables performing multi-threaded work without accidentally reordering results.

Results are delivered as soon as possible, with minimal buffering. Items received out of order are temporarily kept in a binary heap.

Structs

OrderByKey 🔒

Receiver

Receiver that orders messages by an index

RecvError

Error type

SendError

Error type

Sender

A channel sender that orders messages by an index

Enums

SenderKind 🔒

TryRecvError

Error type

TrySendError

Error type

Functions

bounded

Make a blocking channel with finite size

make 🔒

unbounded

Make a channel that can grow until the program runs out of memory